修复手动振动失效的bug
